Revert "Merge branch 'master' of https://github.com/yf225/pytorch into remove_variableimpl"
This reverts commit be454176e6c9d65e7d87977a16dfe9500bdf53ed, reversing
changes made to 46e0a2f18e60c28870861474f2e10d7a4ba27cf9.

+++ caffe2/contrib/aten/aten_op.cc
@@ -14,7 +14,7 @@ OPERATOR_SCHEMA(ATen);
 namespace math {
 template <>
 void Set<at::Half, CPUContext>(
-    const int /*N*/,
+    const size_t /*N*/,
     const at::Half h,
     at::Half* v,
     CPUContext* c) {
